requirements.txt
# Core ML & Embeddings
sentence-transformers==3.4.1
scikit-learn==1.6.1
numpy>=1.26.4
scipy>=1.13.0

# Vector Database
chromadb==0.6.3

# API Framework
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic==2.10.6
aiofiles>=23.2.0

# Data & Utilities
datasets==3.4.0
tqdm==4.67.1
python-dotenv==1.1.0
joblib==1.4.2

# Analysis / Visualization (for clustering script)
matplotlib>=3.7.0
seaborn>=0.12.0
